L.H.M.Scope refactoring
Scope analysis is now always monadic. The monad (State) is used to carry the
symbol table and optionally modify it.

The phantom type parameter of ScopeM is used to control whether changes are made
to the symbol table.

Two type classes (ScopeCheckM and ScopeCheckR) instead of one (ScopeCheck) are
used to distinguish nodes that do and do not introduce bindings.

(This will need to be refined once we support scoped type variables, since they
can be bound almost anywhere.)